# Conflicts:
#	src/modelo/Jugada.java
#	src/modelo/Jugador.java
#	src/modelo/Mano.java
#	test/fiuba/algo3/tests/PruebasJugadaTest.java
#	test/fiuba/algo3/tests/PruebasManoTest.java

+++ src/modelo/ValoresTrucoYEnvido.java
@@ -72,11 +72,11 @@ public class ValoresTrucoYEnvido {
 
 
     public int devolverValorEnvido(int posicion){
-        return (listaValoresTrucoYEnvido.get(posicion -1)).obtenerClave();
+        return (int) (listaValoresTrucoYEnvido.get(posicion -1)).obtenerClave();
     }
 
     public int devolverValorTruco(int posicion){
-        return listaValoresTrucoYEnvido.get(posicion -1).obtenerValor();
+        return (int) listaValoresTrucoYEnvido.get(posicion -1).obtenerValor();
     }
 
     public int devolverTamanio() {
